直接单独生成中文版本
