+++ source/behavioral/_title.swift
@@ -1,13 +1,13 @@
-//: Behavioral |
-//: [Creational](Creational) |
-//: [Structural](Structural)
+//: [行为型模式] |
+//: [创建型模式](Creational) |
+//: [结构型模式](Structural)
 /*:
-Behavioral
-==========
-
->In software engineering, behavioral design patterns are design patterns that identify common communication patterns between objects and realize these patterns. By doing so, these patterns increase flexibility in carrying out this communication.
->
->**Source:** [wikipedia.org](http://en.wikipedia.org/wiki/Behavioral_pattern)
-*/
+ 行为型模式
+ ========
+ 
+ >在软件工程中， 行为型模式为设计模式的一种类型，用来识别对象之间的常用交流模式并加以实现。如此，可在进行这些交流活动时增强弹性。
+ >
+ >**来源：** [维基百科](https://zh.wikipedia.org/wiki/%E8%A1%8C%E7%82%BA%E5%9E%8B%E6%A8%A1%E5%BC%8F)
+ */
 import Swift
 import Foundation
