+++ source/behavioral/iterator.swift
@@ -1,11 +1,9 @@
 /*:
-🍫 Iterator
------------
-
-The iterator pattern is used to provide a standard interface for traversing a collection of items in an aggregate object without the need to understand its underlying structure.
-
-### Example:
-*/
+ 迭代器（Iterator）
+ ---------------
+ 迭代器模式可以让用户通过特定的接口巡访容器中的每一个元素而不用了解底层的实现。
+ ### 示例：
+ */
 struct Novella {
     let name: String
 }
@@ -15,14 +13,14 @@ struct Novellas {
 }
 
 struct NovellasIterator: IteratorProtocol {
-
+    
     private var current = 0
     private let novellas: [Novella]
-
+    
     init(novellas: [Novella]) {
         self.novellas = novellas
     }
-
+    
     mutating func next() -> Novella? {
         defer { current += 1 }
         return novellas.count > current ? novellas[current] : nil
@@ -35,10 +33,9 @@ extension Novellas: Sequence {
     }
 }
 /*:
-### Usage
-*/
-let greatNovellas = Novellas(novellas: [Novella(name: "The Mist")] )
-
+ ### 用法：
+ */
+let greatNovellas = Novellas(novellas: [Novella(name:"红楼梦")])
 for novella in greatNovellas {
-    print("I've read: \(novella)")
+    print("我读了\(novella.name)")
 }
