+++ source/behavioral/mediator.swift
@@ -1,26 +1,10 @@
 /*:
-💐 Mediator
------------
-
-The mediator pattern is used to reduce coupling between classes that communicate with each other. Instead of classes communicating directly, and thus requiring knowledge of their implementation, the classes send messages via a mediator object.
-
-### Example
-*/
-protocol Receiver {
-    associatedtype MessageType
-    func receive(message: MessageType)
-}
-
-protocol Sender {
-    associatedtype MessageType
-    associatedtype ReceiverType: Receiver
-    
-    var recipients: [ReceiverType] { get }
-    
-    func send(message: MessageType)
-}
-
-struct Programmer: Receiver {
+ 中介者（Mediator）
+ ---------------
+ 用一个中介者对象封装一系列的对象交互，中介者使各对象不需要显示地相互作用，从而使耦合松散，而且可以独立地改变它们之间的交互。
+ ### 示例：
+ */
+struct Programmer {
     let name: String
     
     init(name: String) {
@@ -28,12 +12,17 @@ struct Programmer: Receiver {
     }
     
     func receive(message: String) {
-        print("\(name) received: \(message)")
+        print("\(name) 收到消息：\(message)")
     }
 }
 
-final class MessageMediator: Sender {
-    internal var recipients: [Programmer] = []
+protocol MessageSending {
+    func send(message: String)
+}
+
+final class MessageMediator: MessageSending {
+    
+    private var recipients: [Programmer] = []
     
     func add(recipient: Programmer) {
         recipients.append(recipient)
@@ -45,22 +34,21 @@ final class MessageMediator: Sender {
         }
     }
 }
-
 /*:
-### Usage
-*/
-func spamMonster(message: String, worker: MessageMediator) {
+ ### 用法：
+ */
+func spamMonster(message: String, worker: MessageSending) {
     worker.send(message: message)
 }
 
 let messagesMediator = MessageMediator()
 
 let user0 = Programmer(name: "Linus Torvalds")
-let user1 = Programmer(name: "Avadis 'Avie' Tevanian")
+let user1 = Programmer(name: "Dylan Wang")
 messagesMediator.add(recipient: user0)
 messagesMediator.add(recipient: user1)
 
-spamMonster(message: "I'd Like to Add you to My Professional Network", worker: messagesMediator)
+spamMonster(message: "我希望添加您到我的职业网络", worker: messagesMediator)
 /*:
->**Further Examples:** [Design Patterns in Swift](https://github.com/kingreza/Swift-Mediator)
-*/
+ > 更多示例：[Design Patterns in Swift](https://github.com/kingreza/Swift-Mediator)
+ */
