+++ source/behavioral/state.swift
@@ -1,56 +1,56 @@
 /*:
-🐉 State
----------
-
-The state pattern is used to alter the behaviour of an object as its internal state changes.
-The pattern allows the class for an object to apparently change at run-time.
-
-### Example
-*/
+ 状态（State）
+ ---------
+ 在状态模式中，对象的行为是基于它的内部状态而改变的
+ ### 示例：
+ */
 final class Context {
-	private var state: State = UnauthorizedState()
-
+    private var state: State = UnauthorizedState()
+    
     var isAuthorized: Bool {
-        get { return state.isAuthorized(context: self) }
+        get {
+            return state.isAuthorized(context: self)
+        }
     }
-
+    
     var userId: String? {
-        get { return state.userId(context: self) }
+        get {
+            return state.userId(context: self)
+        }
     }
-
-	func changeStateToAuthorized(userId: String) {
-		state = AuthorizedState(userId: userId)
-	}
-
-	func changeStateToUnauthorized() {
-		state = UnauthorizedState()
-	}
-
+    
+    func changeStateToAuthorized(userId: String) {
+        state = AuthorizedState(userId: userId)
+    }
+    
+    func changeStateToUnauthorized() {
+        state = UnauthorizedState()
+    }
 }
 
 protocol State {
-	func isAuthorized(context: Context) -> Bool
-	func userId(context: Context) -> String?
+    func isAuthorized(context: Context) -> Bool
+    func userId(context: Context) -> String?
 }
 
 class UnauthorizedState: State {
-	func isAuthorized(context: Context) -> Bool { return false }
-
-	func userId(context: Context) -> String? { return nil }
+    func isAuthorized(context: Context) -> Bool { return false }
+    
+    func userId(context: Context) -> String? { return nil }
 }
 
 class AuthorizedState: State {
-	let userId: String
-
-	init(userId: String) { self.userId = userId }
-
-	func isAuthorized(context: Context) -> Bool { return true }
-
-	func userId(context: Context) -> String? { return userId }
+    let userId: String
+    
+    init(userId: String) { self.userId = userId }
+    
+    func isAuthorized(context: Context) -> Bool { return true }
+    
+    func userId(context: Context) -> String? { return userId }
 }
 /*:
-### Usage
-*/
+ ### 用法：
+ */
 let userContext = Context()
 (userContext.isAuthorized, userContext.userId)
 userContext.changeStateToAuthorized(userId: "admin")
@@ -58,5 +58,5 @@ userContext.changeStateToAuthorized(userId: "admin")
 userContext.changeStateToUnauthorized()
 (userContext.isAuthorized, userContext.userId)
 /*:
->**Further Examples:** [Design Patterns in Swift](https://github.com/kingreza/Swift-State)
-*/
+ > 更多示例：[Design Patterns in Swift](https://github.com/kingreza/Swift-State)
+ */
