+++ source/behavioral/visitor.swift
@@ -1,20 +1,18 @@
 /*:
-🏃 Visitor
-----------
-
-The visitor pattern is used to separate a relatively complex set of structured data classes from the functionality that may be performed upon the data that they hold.
-
-### Example
-*/
+ 访问者（Visitor）
+ --------------
+ 封装某些作用于某种数据结构中各元素的操作，它可以在不改变数据结构的前提下定义作用于这些元素的新的操作。
+ ### 示例：
+ */
 protocol PlanetVisitor {
-	func visit(planet: PlanetAlderaan)
-	func visit(planet: PlanetCoruscant)
-	func visit(planet: PlanetTatooine)
+    func visit(planet: PlanetAlderaan)
+    func visit(planet: PlanetCoruscant)
+    func visit(planet: PlanetTatooine)
     func visit(planet: MoonJedah)
 }
 
 protocol Planet {
-	func accept(visitor: PlanetVisitor)
+    func accept(visitor: PlanetVisitor)
 }
 
 class MoonJedah: Planet {
@@ -26,36 +24,33 @@ class PlanetAlderaan: Planet {
 }
 
 class PlanetCoruscant: Planet {
-	func accept(visitor: PlanetVisitor) { visitor.visit(planet: self) }
+    func accept(visitor: PlanetVisitor) { visitor.visit(planet: self) }
 }
 
 class PlanetTatooine: Planet {
-	func accept(visitor: PlanetVisitor) { visitor.visit(planet: self) }
+    func accept(visitor: PlanetVisitor) { visitor.visit(planet: self) }
 }
-
-
 
 class NameVisitor: PlanetVisitor {
-	var name = ""
-
-	func visit(planet: PlanetAlderaan)  { name = "Alderaan" }
-	func visit(planet: PlanetCoruscant) { name = "Coruscant" }
-	func visit(planet: PlanetTatooine)  { name = "Tatooine" }
+    var name = ""
+    
+    func visit(planet: PlanetAlderaan)  { name = "Alderaan" }
+    func visit(planet: PlanetCoruscant) { name = "Coruscant" }
+    func visit(planet: PlanetTatooine)  { name = "Tatooine" }
     func visit(planet: MoonJedah)     	{ name = "Jedah" }
 }
-
 /*:
-### Usage
-*/
+ ### 用法：
+ */
 let planets: [Planet] = [PlanetAlderaan(), PlanetCoruscant(), PlanetTatooine(), MoonJedah()]
 
 let names = planets.map { (planet: Planet) -> String in
-	let visitor = NameVisitor()
+    let visitor = NameVisitor()
     planet.accept(visitor: visitor)
-	return visitor.name
+    return visitor.name
 }
 
 names
 /*:
->**Further Examples:** [Design Patterns in Swift](https://github.com/kingreza/Swift-Visitor)
-*/
+ > 更多示例：[Design Patterns in Swift](https://github.com/kingreza/Swift-Visitor)
+ */
