+++ source/creational/_title.swift
@@ -1,13 +1,13 @@
-//: [Behavioral](Behavioral) |
-//: Creational |
-//: [Structural](Structural)
+//: [行为型模式](Behavioral) |
+//: 创建型模式 |
+//: [结构型模式](Structural)
 /*:
-Creational
-==========
-
-> In software engineering, creational design patterns are design patterns that deal with object creation mechanisms, trying to create objects in a manner suitable to the situation. The basic form of object creation could result in design problems or added complexity to the design. Creational design patterns solve this problem by somehow controlling this object creation.
->
->**Source:** [wikipedia.org](http://en.wikipedia.org/wiki/Creational_pattern)
-*/
+ 创建型模式
+ ========
+ 
+ > 创建型模式是处理对象创建的设计模式，试图根据实际情况使用合适的方式创建对象。基本的对象创建方式可能会导致设计上的问题，或增加设计的复杂度。创建型模式通过以某种方式控制对象的创建来解决问题。
+ >
+ >**来源：** [维基百科](https://zh.wikipedia.org/wiki/%E5%89%B5%E5%BB%BA%E5%9E%8B%E6%A8%A1%E5%BC%8F)
+ */
 import Swift
 import Foundation
