+++ source/creational/builder.swift
@@ -1,33 +1,30 @@
 /*:
-👷 Builder
-----------
-
-The builder pattern is used to create complex objects with constituent parts that must be created in the same order or using a specific algorithm. 
-An external class controls the construction algorithm.
-
-### Example
-*/
+ 生成器（Builder）
+ --------------
+ 一种对象构建模式。它可以将复杂对象的建造过程抽象出来（抽象类别），使这个抽象过程的不同实现方法可以构造出不同表现（属性）的对象。
+ ### 示例：
+ */
 class DeathStarBuilder {
-
+    
     var x: Double?
     var y: Double?
     var z: Double?
-
+    
     typealias BuilderClosure = (DeathStarBuilder) -> ()
-
+    
     init(buildClosure: BuilderClosure) {
         buildClosure(self)
     }
 }
 
-struct DeathStar : CustomStringConvertible {
-
+struct DeathStar: CustomStringConvertible {
+    
     let x: Double
     let y: Double
     let z: Double
-
+    
     init?(builder: DeathStarBuilder) {
-
+        
         if let x = builder.x, let y = builder.y, let z = builder.z {
             self.x = x
             self.y = y
@@ -36,21 +33,21 @@ struct DeathStar : CustomStringConvertible {
             return nil
         }
     }
-
-    var description:String {
+    
+    var description: String {
         return "Death Star at (x:\(x) y:\(y) z:\(z))"
     }
 }
 /*:
-### Usage
-*/
+ ### 用法：
+ */
 let empire = DeathStarBuilder { builder in
     builder.x = 0.1
     builder.y = 0.2
     builder.z = 0.3
 }
 
-let deathStar = DeathStar(builder:empire)
+let deathStar = DeathStar(builder: empire)
 /*:
->**Further Examples:** [Design Patterns in Swift](https://github.com/kingreza/Swift-Builder)
-*/
+ > 更多示例：[Design Patterns in Swift](https://github.com/kingreza/Swift-Builder)
+ */
