+++ source/creational/factory.swift
@@ -1,17 +1,15 @@
 /*:
-🏭 Factory Method
------------------
-
-The factory pattern is used to replace class constructors, abstracting the process of object generation so that the type of the object instantiated can be determined at run-time.
-
-### Example
-*/
+ 工厂方法（Factory Method）
+ -----------------------
+ 定义一个创建对象的接口，但让实现这个接口的类来决定实例化哪个类。工厂方法让类的实例化推迟到子类中进行。
+ ### 示例：
+ */
 protocol Currency {
     func symbol() -> String
     func code() -> String
 }
 
-class Euro : Currency {
+class Euro: Currency {
     func symbol() -> String {
         return "€"
     }
@@ -37,22 +35,22 @@ enum Country {
 
 enum CurrencyFactory {
     static func currency(for country:Country) -> Currency? {
-
+        
         switch country {
-            case .spain, .greece :
-                return Euro()
-            case .unitedStates :
-                return UnitedStatesDolar()
-            default:
-                return nil
+        case .spain, .greece :
+            return Euro()
+        case .unitedStates :
+            return UnitedStatesDolar()
+        default:
+            return nil
         }
         
     }
 }
 /*:
-### Usage
-*/
-let noCurrencyCode = "No Currency Code Available"
+ ### 用法：
+ */
+let noCurrencyCode = "无可用货币码"
 
 CurrencyFactory.currency(for: .greece)?.code() ?? noCurrencyCode
 CurrencyFactory.currency(for: .spain)?.code() ?? noCurrencyCode
