+++ source/creational/prototype.swift
@@ -1,27 +1,24 @@
 /*:
-🃏 Prototype
-------------
-
-The prototype pattern is used to instantiate a new object by copying all of the properties of an existing object, creating an independent clone. 
-This practise is particularly useful when the construction of a new object is inefficient.
-
-### Example
-*/
+ 原型（Prototype）
+ --------------
+ 通过“复制”一个已经存在的实例来返回新的实例,而不是新建实例。被复制的实例就是我们所称的“原型”，这个原型是可定制的。
+ ### 示例：
+ */
 class ChungasRevengeDisplay {
     var name: String?
     let font: String
-
+    
     init(font: String) {
         self.font = font
     }
-
+    
     func clone() -> ChungasRevengeDisplay {
-        return ChungasRevengeDisplay(font:self.font)
+        return ChungasRevengeDisplay(font: self.font)
     }
 }
 /*:
-### Usage
-*/
+ ### 用法：
+ */
 let Prototype = ChungasRevengeDisplay(font:"GotanProject")
 
 let Philippe = Prototype.clone()
@@ -33,5 +30,5 @@ Christoph.name = "Christoph"
 let Eduardo = Prototype.clone()
 Eduardo.name = "Eduardo"
 /*:
->**Further Examples:** [Design Patterns in Swift](https://github.com/kingreza/Swift-Prototype)
-*/
+ > 更多示例：[Design Patterns in Swift](https://github.com/kingreza/Swift-Prototype)
+ */
