+++ source/creational/singleton.swift
@@ -1,21 +1,17 @@
 /*:
-💍 Singleton
-------------
-
-The singleton pattern ensures that only one object of a particular class is ever created.
-All further references to objects of the singleton class refer to the same underlying instance.
-There are very few applications, do not overuse this pattern!
-
-### Example:
-*/
+ 单例（Singleton）
+ --------------
+ 单例对象的类必须保证只有一个实例存在。许多时候整个系统只需要拥有一个的全局对象，这样有利于我们协调系统整体的行为
+ ### 示例：
+ */
 class DeathStarSuperlaser {
     static let sharedInstance = DeathStarSuperlaser()
-
+    
     private init() {
-        // Private initialization to ensure just one instance is created.
+        
     }
 }
 /*:
-### Usage:
-*/
+ ### 用法：
+ */
 let laser = DeathStarSuperlaser.sharedInstance
