+++ source/structural/_title.swift
@@ -1,13 +1,13 @@
-//: [Behavioral](Behavioral) |
-//: [Creational](Creational) |
-//: Structural
+//: [行为型模式](Behavioral) |
+//: [创建型模式](Creational) |
+//: 结构型模式
 /*:
-Structural
-==========
-
->In software engineering, structural design patterns are design patterns that ease the design by identifying a simple way to realize relationships between entities.
->
->**Source:** [wikipedia.org](http://en.wikipedia.org/wiki/Structural_pattern)
-*/
+ 结构型模式（Structural）
+ ====================
+ 
+ > 在软件工程中结构型模式是设计模式，借由一以贯之的方式来了解元件间的关系，以简化设计。
+ >
+ >**来源：** [维基百科](https://zh.wikipedia.org/wiki/%E7%B5%90%E6%A7%8B%E5%9E%8B%E6%A8%A1%E5%BC%8F)
+ */
 import Swift
 import Foundation
