+++ source/structural/composite.swift
@@ -1,42 +1,39 @@
 /*:
-🌿 Composite
--------------
-
-The composite pattern is used to create hierarchical, recursive tree structures of related objects where any element of the structure may be accessed and utilised in a standard manner.
-
-### Example
-*/
+ 组合（Composite）
+ --------------
+ 将对象组合成树形结构以表示‘部分-整体’的层次结构。组合模式使得用户对单个对象和组合对象的使用具有一致性。
+ ### 示例：
+ */
 /*:
-Component
-*/
+ 组件（Component）
+ */
 protocol Shape {
     func draw(fillColor: String)
 }
 /*:
-Leafs
-*/
-final class Square : Shape {
+ 叶子节点（Leafs）
+ */
+final class Square: Shape {
     func draw(fillColor: String) {
-        print("Drawing a Square with color \(fillColor)")
+        print("画（\(fillColor)）颜色的方形")
     }
 }
 
-final class Circle : Shape {
+final class Circle: Shape {
     func draw(fillColor: String) {
-        print("Drawing a circle with color \(fillColor)")
+        print("画（\(fillColor)）颜色的圆形")
     }
 }
-
 /*:
-Composite
-*/
-final class Whiteboard : Shape {
+ 组合
+ */
+final class Whiteboard: Shape {
     lazy var shapes = [Shape]()
-
-    init(_ shapes:Shape...) {
+    
+    init(_ shapes: Shape...) {
         self.shapes = shapes
     }
-
+    
     func draw(fillColor: String) {
         for shape in self.shapes {
             shape.draw(fillColor: fillColor)
@@ -44,7 +41,7 @@ final class Whiteboard : Shape {
     }
 }
 /*:
-### Usage:
-*/
+ ### 用法：
+ */
 var whiteboard = Whiteboard(Circle(), Square())
-whiteboard.draw(fillColor: "Red")
+whiteboard.draw(fillColor: "红")
