+++ source/structural/flyweight.swift
@@ -1,9 +1,9 @@
 /*:
-## 🍃 Flyweight
-The flyweight pattern is used to minimize memory usage or computational expenses by sharing as much as possible with other similar objects.
-### Example
-*/
-// Instances of CoffeeFlavour will be the Flyweights
+ 享元（Flyweight）
+ --------------
+ 使用共享物件，用来尽可能减少内存使用量以及分享资讯给尽可能多的相似物件；它适合用于当大量物件只是重复因而导致无法令人接受的使用大量内存。
+ ### 示例：
+ */
 final class SpecialityCoffee: CustomStringConvertible {
     var origin: String
     var description: String {
@@ -11,21 +11,20 @@ final class SpecialityCoffee: CustomStringConvertible {
             return origin
         }
     }
-
+    
     init(origin: String) {
         self.origin = origin
     }
 }
 
-// Menu acts as a factory and cache for CoffeeFlavour flyweight objects
 final class Menu {
     private var coffeeAvailable: [String: SpecialityCoffee] = [:]
-
+    
     func lookup(origin: String) -> SpecialityCoffee? {
         if coffeeAvailable.index(forKey: origin) == nil {
             coffeeAvailable[origin] = SpecialityCoffee(origin: origin)
         }
-
+        
         return coffeeAvailable[origin]
     }
 }
@@ -33,11 +32,11 @@ final class Menu {
 final class CoffeeShop {
     private var orders: [Int: SpecialityCoffee] = [:]
     private var menu = Menu()
-
+    
     func takeOrder(origin: String, table: Int) {
         orders[table] = menu.lookup(origin: origin)
     }
-
+    
     func serve() {
         for (table, origin) in orders {
             print("Serving \(origin) to table \(table)")
@@ -45,8 +44,8 @@ final class CoffeeShop {
     }
 }
 /*:
-### Usage
-*/
+ ### 用法：
+ */
 let coffeeShop = CoffeeShop()
 
 coffeeShop.takeOrder(origin: "Yirgacheffe, Ethiopia", table: 1)
