+++ source/structural/virtual_proxy.swift
@@ -2,7 +2,7 @@
 🍬 Virtual Proxy
 ----------------
 
-The proxy pattern is used to provide a surrogate or placeholder object, which references an underlying object.
+The proxy pattern is used to provide a surrogate or placeholder object, which references an underlying object. 
 Virtual proxy is used for loading object on demand.
 
 ### Example
@@ -13,7 +13,7 @@ protocol HEVSuitMedicalAid {
 
 class HEVSuit : HEVSuitMedicalAid {
     func administerMorphine() -> String {
-        return "Morphine administered."
+        return "Morphine aministered."
     }
 }
