--- source_cn/behavioral/visitor.swift
@@ -1,56 +1,0 @@
-/*:
- 访问者（Visitor）
- --------------
- 封装某些作用于某种数据结构中各元素的操作，它可以在不改变数据结构的前提下定义作用于这些元素的新的操作。
- ### 示例：
- */
-protocol PlanetVisitor {
-    func visit(planet: PlanetAlderaan)
-    func visit(planet: PlanetCoruscant)
-    func visit(planet: PlanetTatooine)
-    func visit(planet: MoonJedah)
-}
-
-protocol Planet {
-    func accept(visitor: PlanetVisitor)
-}
-
-class MoonJedah: Planet {
-    func accept(visitor: PlanetVisitor) { visitor.visit(planet: self) }
-}
-
-class PlanetAlderaan: Planet {
-    func accept(visitor: PlanetVisitor) { visitor.visit(planet: self) }
-}
-
-class PlanetCoruscant: Planet {
-    func accept(visitor: PlanetVisitor) { visitor.visit(planet: self) }
-}
-
-class PlanetTatooine: Planet {
-    func accept(visitor: PlanetVisitor) { visitor.visit(planet: self) }
-}
-
-class NameVisitor: PlanetVisitor {
-    var name = ""
-    
-    func visit(planet: PlanetAlderaan)  { name = "Alderaan" }
-    func visit(planet: PlanetCoruscant) { name = "Coruscant" }
-    func visit(planet: PlanetTatooine)  { name = "Tatooine" }
-    func visit(planet: MoonJedah)     	{ name = "Jedah" }
-}
-/*:
- ### 用法：
- */
-let planets: [Planet] = [PlanetAlderaan(), PlanetCoruscant(), PlanetTatooine(), MoonJedah()]
-
-let names = planets.map { (planet: Planet) -> String in
-    let visitor = NameVisitor()
-    planet.accept(visitor: visitor)
-    return visitor.name
-}
-
-names
-/*:
- > 更多示例：[Design Patterns in Swift](https://github.com/kingreza/Swift-Visitor)
- */
