--- source_cn/creational/factory.swift
@@ -1,58 +1,0 @@
-/*:
- 工厂方法（Factory Method）
- -----------------------
- 定义一个创建对象的接口，但让实现这个接口的类来决定实例化哪个类。工厂方法让类的实例化推迟到子类中进行。
- ### 示例：
- */
-protocol Currency {
-    func symbol() -> String
-    func code() -> String
-}
-
-class Euro: Currency {
-    func symbol() -> String {
-        return "€"
-    }
-    
-    func code() -> String {
-        return "EUR"
-    }
-}
-
-class UnitedStatesDolar : Currency {
-    func symbol() -> String {
-        return "$"
-    }
-    
-    func code() -> String {
-        return "USD"
-    }
-}
-
-enum Country {
-    case unitedStates, spain, uk, greece
-}
-
-enum CurrencyFactory {
-    static func currency(for country:Country) -> Currency? {
-        
-        switch country {
-        case .spain, .greece :
-            return Euro()
-        case .unitedStates :
-            return UnitedStatesDolar()
-        default:
-            return nil
-        }
-        
-    }
-}
-/*:
- ### 用法：
- */
-let noCurrencyCode = "无可用货币码"
-
-CurrencyFactory.currency(for: .greece)?.code() ?? noCurrencyCode
-CurrencyFactory.currency(for: .spain)?.code() ?? noCurrencyCode
-CurrencyFactory.currency(for: .unitedStates)?.code() ?? noCurrencyCode
-CurrencyFactory.currency(for: .uk)?.code() ?? noCurrencyCode
