--- source_cn/creational/prototype.swift
@@ -1,34 +1,0 @@
-/*:
- 原型（Prototype）
- --------------
- 通过“复制”一个已经存在的实例来返回新的实例,而不是新建实例。被复制的实例就是我们所称的“原型”，这个原型是可定制的。
- ### 示例：
- */
-class ChungasRevengeDisplay {
-    var name: String?
-    let font: String
-    
-    init(font: String) {
-        self.font = font
-    }
-    
-    func clone() -> ChungasRevengeDisplay {
-        return ChungasRevengeDisplay(font: self.font)
-    }
-}
-/*:
- ### 用法：
- */
-let Prototype = ChungasRevengeDisplay(font:"GotanProject")
-
-let Philippe = Prototype.clone()
-Philippe.name = "Philippe"
-
-let Christoph = Prototype.clone()
-Christoph.name = "Christoph"
-
-let Eduardo = Prototype.clone()
-Eduardo.name = "Eduardo"
-/*:
- > 更多示例：[Design Patterns in Swift](https://github.com/kingreza/Swift-Prototype)
- */
