--- source_cn/footer.swift
@@ -1,8 +1,0 @@
-/*:
-
-Info
-====
-
-📖 Descriptions from: [Gang of Four Design Patterns Reference Sheet](http://www.blackwasp.co.uk/GangOfFour.aspx)
-
-*/
