--- source_cn/header.swift
@@ -1,16 +1,0 @@
-
-Design Patterns implemented in Swift 3.0
-========================================
-A short cheat-sheet with Xcode 8.2 Playground ([Design-Patterns.playground.zip](https://raw.githubusercontent.com/ochococo/Design-Patterns-In-Swift/master/Design-Patterns.playground.zip)).
-
-👷 Project maintained by: [@nsmeme](http://twitter.com/nsmeme) (Oktawian Chojnacki)
-
-🚀 How to generate README, Playground and zip from source: [GENERATE.md](https://github.com/ochococo/Design-Patterns-In-Swift/blob/master/GENERATE.md)
-
-## Table of Contents
-
-* [Behavioral](#behavioral)
-* [Creational](#creational)
-* [Structural](#structural)
-
-*/
