--- source_cn/structural/composite.swift
@@ -1,47 +1,0 @@
-/*:
- 组合（Composite）
- --------------
- 将对象组合成树形结构以表示‘部分-整体’的层次结构。组合模式使得用户对单个对象和组合对象的使用具有一致性。
- ### 示例：
- */
-/*:
- 组件（Component）
- */
-protocol Shape {
-    func draw(fillColor: String)
-}
-/*:
- 叶子节点（Leafs）
- */
-final class Square: Shape {
-    func draw(fillColor: String) {
-        print("画（\(fillColor)）颜色的方形")
-    }
-}
-
-final class Circle: Shape {
-    func draw(fillColor: String) {
-        print("画（\(fillColor)）颜色的圆形")
-    }
-}
-/*:
- 组合
- */
-final class Whiteboard: Shape {
-    lazy var shapes = [Shape]()
-    
-    init(_ shapes: Shape...) {
-        self.shapes = shapes
-    }
-    
-    func draw(fillColor: String) {
-        for shape in self.shapes {
-            shape.draw(fillColor: fillColor)
-        }
-    }
-}
-/*:
- ### 用法：
- */
-var whiteboard = Whiteboard(Circle(), Square())
-whiteboard.draw(fillColor: "红")
