--- source_cn/structural/decorator.swift
@@ -1,75 +1,0 @@
-/*:
- 修饰（Decorator）
- --------------
- 修饰模式，是面向对象编程领域中，一种动态地往一个类中添加新的行为的设计模式。
- 就功能而言，修饰模式相比生成子类更为灵活，这样可以给某个对象而不是整个类添加一些功能。
- ### 示例：
- */
-protocol Coffee {
-    func getCost() -> Double
-    func getIngredients() -> String
-}
-
-class SimpleCoffee: Coffee {
-    func getCost() -> Double {
-        return 1.0
-    }
-    
-    func getIngredients() -> String {
-        return "Coffee"
-    }
-}
-
-class CoffeeDecorator: Coffee {
-    private let decoratedCoffee: Coffee
-    fileprivate let ingredientSeparator: String = ","
-    
-    required init(decoratedCoffee: Coffee) {
-        self.decoratedCoffee = decoratedCoffee
-    }
-    
-    func getCost() -> Double {
-        return decoratedCoffee.getCost()
-    }
-    
-    func getIngredients() -> String {
-        return decoratedCoffee.getIngredients()
-    }
-}
-
-final class Milk: CoffeeDecorator {
-    required init(decoratedCoffee: Coffee) {
-        super.init(decoratedCoffee: decoratedCoffee)
-    }
-    
-    override func getCost() -> Double {
-        return super.getCost() + 0.5
-    }
-    
-    override func getIngredients() -> String {
-        return super.getIngredients() + ingredientSeparator + "Milk"
-    }
-}
-
-final class WhipCoffee: CoffeeDecorator {
-    required init(decoratedCoffee: Coffee) {
-        super.init(decoratedCoffee: decoratedCoffee)
-    }
-    
-    override func getCost() -> Double {
-        return super.getCost() + 0.7
-    }
-    
-    override func getIngredients() -> String {
-        return super.getIngredients() + ingredientSeparator + "Whip"
-    }
-}
-/*:
- ### 用法：
- */
-var someCoffee: Coffee = SimpleCoffee()
-print("Cost : \(someCoffee.getCost()); Ingredients: \(someCoffee.getIngredients())")
-someCoffee = Milk(decoratedCoffee: someCoffee)
-print("Cost : \(someCoffee.getCost()); Ingredients: \(someCoffee.getIngredients())")
-someCoffee = WhipCoffee(decoratedCoffee: someCoffee)
-print("Cost : \(someCoffee.getCost()); Ingredients: \(someCoffee.getIngredients())")
