--- source_cn/structural/protection_proxy.swift
@@ -1,52 +1,0 @@
-/*:
-☔ Protection Proxy
-------------------
-
-The proxy pattern is used to provide a surrogate or placeholder object, which references an underlying object. 
-Protection proxy is restricting access.
-
-### Example
-*/
-protocol DoorOperator {
-    func open(doors: String) -> String
-}
-
-class HAL9000 : DoorOperator {
-    func open(doors: String) -> String {
-        return ("HAL9000: Affirmative, Dave. I read you. Opened \(doors).")
-    }
-}
-
-class CurrentComputer : DoorOperator {
-    private var computer: HAL9000!
-
-    func authenticate(password: String) -> Bool {
-
-        guard password == "pass" else {
-            return false;
-        }
-
-        computer = HAL9000()
-
-        return true
-    }
-
-    func open(doors: String) -> String {
-
-        guard computer != nil else {
-            return "Access Denied. I'm afraid I can't do that."
-        }
-
-        return computer.open(doors: doors)
-    }
-}
-/*:
-### Usage
-*/
-let computer = CurrentComputer()
-let podBay = "Pod Bay Doors"
-
-computer.open(doors: podBay)
-
-computer.authenticate(password: "pass")
-computer.open(doors: podBay)
